Client/server resource split
