+++ src/client/java/tk/estecka/invarpaint/mixin/client/ItemMixin.java
@@ -1,4 +1,4 @@
-package tk.estecka.invarpaint.mixin;
+package tk.estecka.invarpaint.mixin.client;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
